refactor and pytest trial
