+++ dags/logs/scheduler/latest
@@ -1,1 +1,1 @@
-/opt/airflow/logs/scheduler/2023-06-26
+/opt/airflow/logs/scheduler/2023-06-27
